Update Timetable variable type for printSchedule method

+++ src/main/java/seedu/sherpass/util/Timetable.java
@@ -159,9 +159,9 @@ public class Timetable {
     private void printSchedule() {
         String day = localDate.format(dayOnlyFormat);
         String date = localDate.format(dateOnlyFormat);
-        int taskLength = findTaskLength(tasks);
-        int doOnDateLength = DATE_SPACE_FULL_LENGTH;
-        int partitionLength = calcPartitionLength(taskLength, doOnDateLength);
+        long taskLength = findTaskLength(tasks);
+        long doOnDateLength = DATE_SPACE_FULL_LENGTH;
+        long partitionLength = calcPartitionLength(taskLength, doOnDateLength);
 
         if (!tasks.isEmpty()) {
             printTimetable(day, date, tasks, ui, taskLength, doOnDateLength, partitionLength);
